extract all export/record-related code to a separate file/class
Also add stub 4konv project.
The code currently doesn't even build. I will be resolving issues
progressively as I dive in.
There are also a bit of changes to the project files to allow it to
build on msvs 2019.

The goal here is to have a separate faster-than-realtime
renoise-to-h/inc converter utility to allow for faster iteration time.